+++ src/compiler/exp_desc.rs
@@ -27,6 +27,8 @@ pub(super) enum PlaceExp {
     Local(u8),
     /// A global variable, and its index in the list of string literals
     Global(u8),
+    /// An upvalue (captured outer variable), and its index in the upvalue list
+    Upvalue(u8),
     /// A table index, with `[` and `]`
     TableIndex,
     /// A field access, and the index of the field's identifier in the list of